Fix bug 2847758: "operator!=(lhs,scr) sometimes wrongly returns true".
In operator!= (const Multivector<Scalar_T,LO,HI>& lhs, const Scalar_T& scr).,
with GluCat 0.4.1 and earlier, when lhs is of type matrix_multi_t,
it is possible that lhs.frame().count() !=0 and yet the matrix is scr times the identity matrix.
In this case operator!=() returns false when it should return true.