Update to version 0.5.0.
Options for framed_multi map: Make these more sensible:
Change options "--with-hash-map" and "--with-unordered-map" to
"--with-map=(map|hash|unordered)". Make "unordered" the default value.
When "hash" is selected, check and add compiler flag "-Wno-deprecated",
check header files <ext/hash_map> and define _GLUCAT_USE_GNU_CXX_HASH_MAP.
If <backward/hash_map> is also found, define _GLUCAT_USE_BACKWARD_HASH_MAP.

Options for random number generation:
Add new option "--with-random=(tr1|gsl)", with "tr1" the default value.
This selects the type of random number generator to use.
If "gsl" was selected, use AC_CHECK_LIB to check for libgsl.a and libgslcblas.a,
if found, check header file <gsl/gsl_rng.h>, and if that is found, define
_GLUCAT_USE_GSL_RANDOM and append "-lgsl -lgslcblas" to the variable
all_libraries. If "tr1" was selected, or in the default case, check
headr file <tr1/random>, but do not define any macros.

Option for PyCliCal:
Add new option "--enable-pyclical". This uses Cython to build the PyCliCal
prototype. If "--disable-pyclical" was selected, or in the default case,
append "pyclical" to the variable DO_NOT_COMPILE. If "-enable-pyclical" was
selected, use AM_CONDITIONAL to set the condition make_pyclical.

Compilation with Intel C++:
For icpc version 11 or greater, check and add compiler flag "-fp-model precise"
rather than "-mp1".

Option for QD: Make this more consistent with others:
If "--with-qd" was selected, use AC_CHECK_LIB to check for libqd.a,
if found, check header file <qd/qd_config.h>, and if that is found, define
_GLUCAT_USE_QD and append "-lqd" to the variable all_libraries.

Do not append USER_INCLUDES to all_includes twice.
Do not append USER_LDFLAGS to all_libraries twice.

Move parentheses on a separate line back to the end of the previous line.