This script is now more usable as a demo. Many changes have been made, including:
o New arguments:
- azimuth : Angle about a vertical axis used to define the viewpoint.
- jitter : Angle in degrees to use for each step in the rotation.

o More sensible default values for arguments.

o Do not use the twisted adjoint action |=, but instead calculate ir, the inverse of the rotor r once, and use the adjoint action u = r * u * ir. This works because r must be even, since it is th exponential of a bivector.

o Do not plot lines. The code is now fast enough that this is no longer useful, since the hull of the scattered points can now more easily be seen.

o Do not use plt.show(). This caused the plots to corrupt when the next figure was started.

o Ensure the points lie within the arena by doubling the arena width and  recolouring existing points when necessary.