Include header dependencies.
Use optional preprocessor symbols to control inclusions in this file.

Use "#if defined()" instead of "#ifdef".

Declare exp(): exponential of multivector, and make it a friend of framed_multi<>.

Change hash() to index_set_hash().

Define hash_size_t.

Declare a private constructor that uses a hash size.
This is to try to optimize the allocation for the hash and unordered_map implementations of framed_multi<>.

Declare a new member function nbr_terms(): number of terms.