Include header dependencies.
Use "#if defined()" instead of "#ifdef".

Rewrite basis_element() so that the basis_cache uses pointers to matrices instead of matrices.
Make the use of the basis_cache depend on the actual frame count().

Overhaul the implementation of sqrt() and log() so that the Pade' approximation depends on the numeric precision.

Implement exp(): exponential of multivector, using clifford_exp().

Cosmetic: Use traits_t consistently.